# ProjetMinuteMed/src/main/webapp/swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Gives all the details of the MinuteMed APIs",
    "version" : "1.0.0",
    "title" : "MinuteMedAPI"
  },
  "host" : "localhost:8080",
  "basePath" : "/api",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/nodes/APHP/{idAPHP}/hospitals/{idHospital}/poles/{idPole}/services/{idService}" : {
      "get" : {
        "tags" : [ "Arborescence REST Endpoint" ],
        "summary" : "Get a specific Service.",
        "description" : "",
        "operationId" : "getService",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idAPHP",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "idHospital",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "idPole",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "idService",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/NodeService"
            }
          },
          "201" : {
            "description" : "The Service is returned."
          },
          "400" : {
            "description" : "Invalid input"
          },
          "404" : {
            "description" : "Service not found"
          }
        }
      }
    },
...
    "/records" : {
      "get" : {
        "tags" : [ "MedicalRecords REST Endpoint" ],
        "summary" : "Get all Medical Records.",
        "description" : "",
        "operationId" : "getAllMedicalRecord",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/MedicalRecordDTO"
            }
          },
          "201" : {
            "description" : "All the Medical Records are returned."
          }
        }
      },
      "post" : {
        "tags" : [ "MedicalRecords REST Endpoint" ],
        "summary" : "Create a medical record",
        "description" : "",
        "operationId" : "createMedicalRecord",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "responses" : {
          "201" : {
            "description" : "The Medical Record is created !"
          },
          "400" : {
            "description" : "Invalid input"
          }
        }
      }
    },
    "/records/{idRecord}" : {
      "get" : {
        "tags" : [ "MedicalRecords REST Endpoint" ],
        "summary" : "Get a specific Medical Record.",
        "description" : "",
        "operationId" : "getMedicalRecord",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idRecord",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/MedicalRecordDTO"
            }
          },
          "201" : {
            "description" : "The Medical Record is returned."
          },
          "400" : {
            "description" : "Invalid input"
          },
          "404" : {
            "description" : "Medical Record not found"
          }
        }
      },
      "delete" : {
        "tags" : [ "MedicalRecords REST Endpoint" ],
        "summary" : "Delete a Medical Record.",
        "description" : "",
        "operationId" : "deleteMedicalRecord",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idRecord",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "201" : {
            "description" : "The Medical Record is deleted."
          },
          "400" : {
            "description" : "Invalid input"
          },
          "404" : {
            "description" : "Medical Record not found"
          }
        }
      }
    },
    "/records/{idRecord}/diagnostics" : {
      "get" : {
        "tags" : [ "MedicalRecords REST Endpoint" ],
        "summary" : "Get all Diagnostics associated with a Medical Record.",
        "description" : "",
        "operationId" : "getDiagnostics",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idRecord",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/DiagnosticDTO"
              }
            }
          },
          "201" : {
            "description" : "All the Diagnostics are returned."
          },
          "400" : {
            "description" : "Invalid input"
          },
          "404" : {
            "description" : "Medical Record not found"
          }
        }
      },
      "post" : {
        "tags" : [ "MedicalRecords REST Endpoint" ],
        "summary" : "Create a Diagnostic in the Medical Record associated",
        "description" : "",
        "operationId" : "createDiagnostic",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idRecord",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "201" : {
            "description" : "The Diagnostic is created !"
          },
          "400" : {
            "description" : "Invalid input"
          }
        }
      }
    },
    "/records/{idRecord}/diagnostics/{idDiagnostic}" : {
      "get" : {
        "tags" : [ "MedicalRecords REST Endpoint" ],
        "summary" : "Get a specific Diagnostic associated with a Medical Record.",
        "description" : "",
        "operationId" : "getDiagnostic",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idRecord",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "idDiagnostic",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/DiagnosticDTO"
            }
          },
          "201" : {
            "description" : "The Diagnostic is returned."
          },
          "400" : {
            "description" : "Invalid input"
          },
          "404" : {
            "description" : "Medical Record or Diagnostic not found"
          }
        }
      },
      "delete" : {
        "tags" : [ "MedicalRecords REST Endpoint" ],
        "summary" : "Delete a Diagnostic.",
        "description" : "",
        "operationId" : "deleteDiagnostic",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idRecord",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "idDiagnostic",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "201" : {
            "description" : "The Diagnostic is deleted."
          },
          "400" : {
            "description" : "Invalid input"
          },
          "404" : {
            "description" : "Diagnostic not found"
          }
        }
      }
    },
    "/records/{idRecord}/dosages" : {
      "get" : {
        "tags" : [ "MedicalRecords REST Endpoint" ],
        "summary" : "Get all Dosages associated with a Medical Record.",
        "description" : "",
        "operationId" : "getAllDosages",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idRecord",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/DosageDTO"
              }
            }
          },
          "201" : {
            "description" : "All the Dosages are returned."
          },
          "400" : {
            "description" : "Invalid input"
          },
          "404" : {
            "description" : "Medical Record not found"
          }
        }
      },
      "post" : {
        "tags" : [ "MedicalRecords REST Endpoint" ],
        "summary" : "Create a Dosage in the Medical Record associated",
        "description" : "",
        "operationId" : "createDosage",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idRecord",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "201" : {
            "description" : "The Dosage is created !"
          },
          "400" : {
            "description" : "Invalid input"
          }
        }
      }
    },
    "/records/{idRecord}/dosages/{idDosage}" : {
      "get" : {
        "tags" : [ "MedicalRecords REST Endpoint" ],
        "summary" : "Get a specific Dosage associated with a Medical Record.",
        "description" : "",
        "operationId" : "getDosage",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idRecord",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "idDosage",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/DosageDTO"
            }
          },
          "201" : {
            "description" : "The Dosage is returned."
          },
          "400" : {
            "description" : "Invalid input"
          },
          "404" : {
            "description" : "Medical Record or Dosage not found"
          }
        }
      },
      "delete" : {
        "tags" : [ "MedicalRecords REST Endpoint" ],
        "summary" : "Delete a Dosage.",
        "description" : "",
        "operationId" : "deleteDosage",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idRecord",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "idDosage",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "201" : {
            "description" : "The Dosage is deleted."
          },
          "400" : {
            "description" : "Invalid input"
          },
          "404" : {
            "description" : "Dosage not found"
          }
        }
      }
    },
    "/records/{idRecord}/exams" : {
      "get" : {
        "tags" : [ "MedicalRecords REST Endpoint" ],
        "summary" : "Get all Exams associated with a Medical Record.",
        "description" : "",
        "operationId" : "getExams",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idRecord",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ExamDTO"
              }
            }
          },
          "201" : {
            "description" : "All the Exams are returned."
          },
          "400" : {
            "description" : "Invalid input"
          },
          "404" : {
            "description" : "Medical Record not found"
          }
        }
      },
      "post" : {
        "tags" : [ "MedicalRecords REST Endpoint" ],
        "summary" : "Create an Exam in the Medical Record associated",
        "description" : "",
        "operationId" : "createExam",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idRecord",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "201" : {
            "description" : "The Exam is created !"
          },
          "400" : {
            "description" : "Invalid input"
          }
        }
      }
    },
    "/records/{idRecord}/exams/{idExam}" : {
      "get" : {
        "tags" : [ "MedicalRecords REST Endpoint" ],
        "summary" : "Get a specific Exam associated with a Medical Record.",
        "description" : "",
        "operationId" : "getExam",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idRecord",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "idExam",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ExamDTO"
            }
          },
          "201" : {
            "description" : "The Exam is returned."
          },
          "400" : {
            "description" : "Invalid input"
          },
          "404" : {
            "description" : "Medical Record or Exam not found"
          }
        }
      },
      "delete" : {
        "tags" : [ "MedicalRecords REST Endpoint" ],
        "summary" : "Delete an Exam.",
        "description" : "",
        "operationId" : "deleteExam",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "idRecord",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "idExam",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "201" : {
            "description" : "The Exam is deleted."
          },
          "400" : {
            "description" : "Invalid input"
          },
          "404" : {
            "description" : "Exam not found"
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "DiagnosticDTO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "creatorId" : {
          "type" : "integer",
          "format" : "int64"
        },
        "medicalRecordId" : {
          "type" : "integer",
          "format" : "int64"
        },
        "title" : {
          "type" : "string"
        },
        "creationDate" : {
          "type" : "string"
        },
        "body" : {
          "type" : "string"
        },
        "draft" : {
          "type" : "boolean",
          "default" : false
        }
      }
    },
...
    "ExamDTO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "doctorId" : {
          "type" : "integer",
          "format" : "int64"
        },
        "medicalRecordId" : {
          "type" : "integer",
          "format" : "int64"
        },
        "title" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "dateExam" : {
          "type" : "string"
        },
        "resultExam" : {
          "$ref" : "#/definitions/ResultExamDTO"
        },
        "draft" : {
          "type" : "boolean",
          "default" : false
        }
      }
    },
    "MedicalRecordDTO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "ss" : {
          "type" : "string"
        },
        "firstName" : {
          "type" : "string"
        },
        "lastName" : {
          "type" : "string"
        },
        "address" : {
          "type" : "string"
        },
        "birthday" : {
          "type" : "string"
        },
        "gender" : {
          "type" : "string"
        },
        "phoneNumber" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "careUnit" : {
          "$ref" : "#/definitions/NodeDTO"
        }
      }
    },
...
    "NodeService" : {
      "type" : "object",
      "required" : [ "father", "floor", "name" ],
      "properties" : {
        "idNode" : {
          "type" : "integer",
          "format" : "int64"
        },
        "floor" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "father" : {
          "$ref" : "#/definitions/NodePole"
        }
      }
    },
...
  }
}